+++ CHANGELOG.md
@@ -5,9 +5,25 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased](https://github.com/global-121/121-platform/compare/v1.95.0...master)
+## [Unreleased](https://github.com/global-121/121-platform/compare/v1.96.0...master)
 
 ---
+
+## [1.96.0](https://github.com/global-121/121-platform/compare/v1.95.1...v1.96.0) - 2023-08-16
+
+### Fixed
+
+- Open voucher from registration page
+
+### Changed
+
+- Moved (CI/Lint) tests from Azure Pipelines to GitHub Actions
+
+## [1.95.1](https://github.com/global-121/121-platform/compare/v1.95.0...v1.95.1) - 2023-08-14
+
+### Added
+
+- Added 1 to 1 relation between Twilio message and transaction to improve performance of handling status updates
 
 ## [1.95.0](https://github.com/global-121/121-platform/compare/v1.94.1...v1.95.0) - 2023-08-11